requirements.txt
flask
python-dotenv
deepl
Werkzeug
